requirements.txt
autogen==0.3.1
pandas==2.1.1
numpy==1.25.2
python-dotenv==1.0.0
groq==0.4.0
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.18.0
scikit-learn==1.5.2
openpyxl==3.1.2
python-docx==0.8.11

